% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puerto_rico.R
\docType{data}
\name{puerto_rico}
\alias{puerto_rico}
\title{Shape file for the country of Puerto Rico}
\format{A shape file with a geometry type of multipolygon with dimensions of XY.}
\source{
\url{https://gadm.org/download_country_v3.html}
}
\usage{
puerto_rico
}
\description{
A shape file that contains the location information of polygons
that make up Puerto Rico.
}
\keyword{datasets}
