% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_1.R
\name{team_1}
\alias{team_1}
\title{Impletement the Team 1 Solution in Lab 2}
\usage{
team_1(file, tolerance = 0.1)
}
\arguments{
\item{file}{A path to the .shp file. Or a loaded sf shape file created using sf
package with a geometry type of mutipolygon. If a loaded sf shape is supplied, the
the tolerance is turned off.}

\item{tolerance}{The value used to thin the data.}
}
\value{
A data.frame with the geometry information extracted to the top level.
All other columns are kept. The new columns are as follows:

\tabular{ll}{
 listno\tab The orginal rows that the expanded data coming from.\cr
 long\tab Longitude. \cr
 lat\tab Latitude. \cr
 group\tab The group used to draw polygen. \cr
 order\tab The order for the points within group. }
}
\description{
This function need a file path as an input to read the .shp file, and it will
extract the geometry information to the top level.
}
\examples{

# A shape file for Puerto Rico is already stored in the package
puerto_rico

# Use the team_1 function to convert the shape file to a dataframe
puerto_rico_df <- team_1(puerto_rico)
head(puerto_rico_df)

# Create a plot of Puerto Rico using the dataframe
library(ggplot2)
ggplot(puerto_rico_df, aes(x = long, y = lat, group = group)) + geom_polygon()
}
