% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_5.R
\name{team_5}
\alias{team_5}
\title{Team 5 Function}
\usage{
team_5(file, tolerance = 0.1)
}
\arguments{
\item{file}{This is the shape file that will be converted to a datafarme. It can either
be a file path in the form of a character string (ends in .shp) or a sf shape file
object created using the sf package with a geometry type of multipolygon.}

\item{tolerance}{This controls how much the shape file is thinned. The larger it is made
the less detail the shape file will have. The tolerance is set to 0.1 by default.}
}
\value{
A dataframe created from the shape file. See the details for the
  variables included in the dataframe.
}
\description{
Function based on the work by lab group 5 that converts a country shape file into a dataframe.
}
\details{
The variables included in the dataframe that is returned from \code{team_5}
are as follows.
\itemize{
\item country = country depicted by the data
\item group = indicates which polygon a set of points corresponds to
\item order = order in which the points in a polygon should be arranged
\item long = longitude of the point
\item lat = latitude of the point
}
}
\examples{

# A shape file for Puerto Rico is already stored in the package
puerto_rico

# Use the team_5 function to convert the shape file to a dataframe
puerto_rico_df <- team_5(puerto_rico, 0.001)
head(puerto_rico_df)

# Create a plot of Puerto Rico using the dataframe
library(ggplot2)
ggplot(puerto_rico_df, aes(x = long, y = lat, group = group)) + geom_polygon()
}
