% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{team_7}
\alias{team_7}
\title{Team 7}
\usage{
team_7(file, tolerance = 0.1)
}
\arguments{
\item{file}{This is a file path to a shape file.}

\item{tolerance}{is the value used for thinning the polygon. The default of this value was set as 0.1.}
}
\value{
A dataframe created from the shape file.
}
\description{
This function was based on the work of team 7 in lab 2. From lab 2, this team convert a shp file from Austrailia to a data file.
Our goal is to wrap the solution from this team into a function.
}
\details{
The variables included in the dataframe that is returned from \code{team_5}
are as follows.
}
\examples{

# A shape file for Puerto Rico is already stored in the package
puerto_rico

# Use the team_7 function to convert the shape file to a dataframe
puerto_rico_df <- team_7(puerto_rico, 0.001)
head(puerto_rico_df)

# Create a plot of Puerto Rico using the dataframe
library(ggplot2)
ggplot(puerto_rico_df, aes(x = long, y = lat, group = group)) + geom_polygon()
}
